consumables page refactored to new rule card structure
