+++ phb-app/src/components/RuleCards/index.js
@@ -3,11 +3,21 @@ import ArmorCustomization from './ArmorCustomization.vue';
 import ArmorPieces from './ArmorPieces.vue';
 import ArmorProficiency from './ArmorProficiency.vue';
 import CreatingNewArmor from './CreatingNewArmor.vue';
+import Grenades from './Grenades.vue';
+import HeavyWeaponCharges from './HeavyWeaponCharges.vue';
+import MediGel from './MediGel.vue';
+import ThermalClips from './ThermalClips.vue';
+import ThermalClipsVariant from './ThermalClipsVariant.vue';
 
 export default {
   ArmorOverview,
   ArmorCustomization,
   ArmorPieces,
   ArmorProficiency,
-  CreatingNewArmor
+  CreatingNewArmor,
+  Grenades,
+  HeavyWeaponCharges,
+  MediGel,
+  ThermalClips,
+  ThermalClipsVariant
 };
